+++ tests/api/utils/test_definition_rendering.py
@@ -21,7 +21,9 @@ from unittest import mock
 import pytest
 from snowflake.cli.api.exceptions import CycleDetectedError, InvalidTemplate
 from snowflake.cli.api.project.definition import load_project
+from snowflake.cli.api.project.errors import SchemaValidationError
 from snowflake.cli.api.utils.definition_rendering import render_definition_template
+from snowflake.cli.api.utils.models import ProjectEnvironment
 
 
 @mock.patch.dict(os.environ, {}, clear=True)
@@ -35,11 +37,15 @@ def test_resolve_variables_in_project_no_cross_variable_dependencies():
         },
     }
 
-    result = render_definition_template(definition)
+    result = render_definition_template(definition, {}).project_context
 
     assert result == {
-        "definition_version": "1.1",
-        "env": {"number": 1, "string": "foo", "boolean": True},
+        "ctx": {
+            "definition_version": "1.1",
+            "env": ProjectEnvironment(
+                {"number": 1, "string": "foo", "boolean": True}, {}
+            ),
+        }
     }
 
 
@@ -53,34 +59,28 @@ def test_resolve_variables_in_project_cross_variable_dependencies():
             "C": "<% ctx.env.B %> and <% ctx.env.A %>",
         },
     }
-    result = render_definition_template(definition)
+    result = render_definition_template(definition, {}).project_context
 
     assert result == {
-        "definition_version": "1.1",
-        "env": {"A": 42, "B": "b=42", "C": "b=42 and 42"},
+        "ctx": {
+            "definition_version": "1.1",
+            "env": ProjectEnvironment({"A": 42, "B": "b=42", "C": "b=42 and 42"}, {}),
+        }
     }
 
 
 @mock.patch.dict(os.environ, {}, clear=True)
-def test_no_resolve_in_version_1_0():
+def test_no_resolve_in_version_1():
     definition = {
-        "definition_version": "1.0",
+        "definition_version": "1",
         "env": {
-            "A": 42,
+            "A": "42",
             "B": "b=<% ctx.env.A %>",
             "C": "<% ctx.env.B %> and <% ctx.env.A %>",
         },
     }
-    result = render_definition_template(definition)
-
-    assert result == {
-        "definition_version": "1.0",
-        "env": {
-            "A": 42,
-            "B": "b=<% ctx.env.A %>",
-            "C": "<% ctx.env.B %> and <% ctx.env.A %>",
-        },
-    }
+    with pytest.raises(SchemaValidationError):
+        render_definition_template(definition, {})
 
 
 @mock.patch.dict(os.environ, {}, clear=True)
@@ -90,13 +90,18 @@ def test_resolve_variables_in_project_cross_project_dependencies():
         "streamlit": {"name": "my_app"},
         "env": {"app": "name of streamlit is <% ctx.streamlit.name %>"},
     }
-    result = render_definition_template(definition)
+    result = render_definition_template(definition, {}).project_context
     assert result == {
-        "definition_version": "1.1",
-        "streamlit": {"name": "my_app"},
-        "env": {
-            "app": "name of streamlit is my_app",
-        },
+        "ctx": {
+            "definition_version": "1.1",
+            "streamlit": {"name": "my_app"},
+            "env": ProjectEnvironment(
+                {
+                    "app": "name of streamlit is my_app",
+                },
+                {},
+            ),
+        }
     }
 
 
@@ -119,46 +124,55 @@ def test_resolve_variables_in_project_environment_variables_precedence():
             "test_variable_2": "<% ctx.env.value_from_env %>",
         },
     }
-    result = render_definition_template(definition)
+    result = render_definition_template(definition, {}).project_context
 
     assert result == {
-        "definition_version": "1.1",
-        "env": {
-            "should_be_replaced_by_env": "test failed",
-            "test_variable": "new_lowercase_value and new_uppercase_value",
-            "test_variable_2": "this comes from os.environ",
-        },
+        "ctx": {
+            "definition_version": "1.1",
+            "env": ProjectEnvironment(
+                {
+                    "should_be_replaced_by_env": "test failed",
+                    "test_variable": "new_lowercase_value and new_uppercase_value",
+                    "test_variable_2": "this comes from os.environ",
+                },
+                {},
+            ),
+        }
     }
-    assert result["env"]["lowercase"] == "new_lowercase_value"
-    assert result["env"]["UPPERCASE"] == "new_uppercase_value"
-    assert result["env"]["should_be_replaced_by_env"] == "test succeeded"
-    assert result["env"]["value_from_env"] == "this comes from os.environ"
+    assert result["ctx"]["env"]["lowercase"] == "new_lowercase_value"
+    assert result["ctx"]["env"]["UPPERCASE"] == "new_uppercase_value"
+    assert result["ctx"]["env"]["should_be_replaced_by_env"] == "test succeeded"
+    assert result["ctx"]["env"]["value_from_env"] == "this comes from os.environ"
 
 
 @mock.patch.dict(os.environ, {"env_var": "<% ctx.definition_version %>"}, clear=True)
 def test_env_variables_do_not_get_resolved():
     definition = {
         "definition_version": "1.1",
-        "native_app": {
-            "name": "test_source_<% ctx.env.env_var %>",
-        },
+        "native_app": {"name": "test_source_<% ctx.env.env_var %>", "artifacts": []},
         "env": {
             "reference_to_name": "<% ctx.native_app.name %>",
         },
     }
-    result = render_definition_template(definition)
+    result = render_definition_template(definition, {}).project_context
 
     assert result == {
-        "definition_version": "1.1",
-        "native_app": {
-            "name": "test_source_<% ctx.definition_version %>",
-        },
-        "env": {
-            "reference_to_name": "test_source_<% ctx.definition_version %>",
-        },
+        "ctx": {
+            "definition_version": "1.1",
+            "native_app": {
+                "name": "test_source_<% ctx.definition_version %>",
+                "artifacts": [],
+            },
+            "env": ProjectEnvironment(
+                {
+                    "reference_to_name": "test_source_<% ctx.definition_version %>",
+                },
+                {},
+            ),
+        }
     }
 
-    assert result["env"]["env_var"] == "<% ctx.definition_version %>"
+    assert result["ctx"]["env"]["env_var"] == "<% ctx.definition_version %>"
 
 
 @pytest.mark.parametrize(
@@ -180,17 +194,18 @@ def test_env_variables_do_not_get_resolved():
         },
         {
             "definition_version": "1.1",
-            "native_app": {"name": "test_<% ctx.env.A %>"},
+            "native_app": {"name": "test_<% ctx.env.A %>", "artifacts": []},
             "env": {"A": "<% ctx.native_app.name %>"},
         },
         {
             "definition_version": "1.1",
-            "native_app": {"name": "test_<% ctx.native_app.name %>"},
+            "native_app": {"name": "test_<% ctx.native_app.name %>", "artifacts": []},
         },
         {
             "definition_version": "1.1",
             "native_app": {
                 "name": "test_<% ctx.native_app.source_stage %>",
+                "artifacts": [],
                 "source_stage": "stage <% ctx.native_app.name %>",
             },
         },
@@ -198,7 +213,7 @@ def test_env_variables_do_not_get_resolved():
 )
 def test_resolve_variables_error_on_cycle(definition):
     with pytest.raises(CycleDetectedError) as err:
-        render_definition_template(definition)
+        render_definition_template(definition, {})
 
     assert err.value.message.startswith("Cycle detected in templating variable ")
 
@@ -232,11 +247,11 @@ def test_resolve_variables_error_on_cycle(definition):
 )
 def test_resolve_variables_reference_non_scalar(definition, error_var):
     with pytest.raises(InvalidTemplate) as err:
-        render_definition_template(definition)
+        render_definition_template(definition, {})
 
     assert (
         err.value.message
-        == f"Template variable {error_var} does not contain a valid value"
+        == f"Template variable {error_var} does not have a scalar value"
     )
 
 
@@ -246,26 +261,29 @@ def test_resolve_variables_blank_is_ok():
         "definition_version": "1.1",
         "native_app": {
             "name": "<% ctx.env.blank_default_env %>",
-            "source_stage": "",
             "deploy_root": "<% ctx.env.blank_env %>",
+            "artifacts": [],
         },
         "env": {
             "blank_default_env": "",
-            "refers_to_blank": "<% ctx.native_app.source_stage %>",
         },
     }
-    result = render_definition_template(definition)
+    result = render_definition_template(definition, {}).project_context
 
     assert result == {
-        "definition_version": "1.1",
-        "native_app": {"name": "", "source_stage": "", "deploy_root": ""},
-        "env": {
-            "blank_default_env": "",
-            "refers_to_blank": "",
-        },
+        "ctx": {
+            "definition_version": "1.1",
+            "native_app": {"name": "", "deploy_root": "", "artifacts": []},
+            "env": ProjectEnvironment(
+                {
+                    "blank_default_env": "",
+                },
+                {},
+            ),
+        }
     }
 
-    assert result["env"]["blank_env"] == ""
+    assert result["ctx"]["env"]["blank_env"] == ""
 
 
 @pytest.mark.parametrize(
@@ -277,16 +295,30 @@ def test_resolve_variables_blank_is_ok():
             "Could not find template variable ctx.streamlit.name",
         ),
         ({"app": "<% ctx.foo %>"}, "Could not find template variable ctx.foo"),
+        ({"app": "<% ctx.env.get %>"}, "Could not find template variable ctx.env.get"),
+        (
+            {"app": "<% ctx.env.__class__ %>"},
+            "Could not find template variable ctx.env.__class__",
+        ),
+        (
+            {"app": "<% ctx.native_app.__class__ %>"},
+            "Could not find template variable ctx.native_app.__class__",
+        ),
+        (
+            {"app": "<% ctx.native_app.schema %>"},
+            "Could not find template variable ctx.native_app.schema",
+        ),
     ],
 )
 def test_resolve_variables_fails_if_referencing_unknown_variable(env, msg):
     definition = {
         "definition_version": "1.1",
+        "native_app": {"name": "test_name", "artifacts": []},
         "env": env,
     }
     with pytest.raises(InvalidTemplate) as err:
-        render_definition_template(definition)
-    assert msg in str(err.value)
+        render_definition_template(definition, {})
+    assert msg == err.value.message
 
 
 @mock.patch.dict(os.environ, {}, clear=True)
@@ -310,13 +342,16 @@ def test_unquoted_template_usage_in_strings_yaml():
         p.write_text(dedent(text))
         project_definition = load_project([p]).project_definition
 
-    assert project_definition.env == {
-        "block_multiline": "this is multiline string \nwith template Snowflake is great!\n",
-        "flow_multiline_not_quoted": "this is multiline string with template Snowflake is great!",
-        "flow_multiline_quoted": "this is multiline string with template Snowflake is great!",
-        "single_line": "Snowflake is great!",
-        "value": "Snowflake is great!",
-    }
+    assert project_definition.env == ProjectEnvironment(
+        {
+            "block_multiline": "this is multiline string \nwith template Snowflake is great!\n",
+            "flow_multiline_not_quoted": "this is multiline string with template Snowflake is great!",
+            "flow_multiline_quoted": "this is multiline string with template Snowflake is great!",
+            "single_line": "Snowflake is great!",
+            "value": "Snowflake is great!",
+        },
+        {},
+    )
 
 
 @mock.patch.dict(os.environ, {"var_with_yml": "    - app/*\n    - src\n"}, clear=True)
@@ -327,16 +362,21 @@ def test_injected_yml_in_env_should_not_be_expanded():
             "test_env": "<% ctx.env.var_with_yml %>",
         },
     }
-    result = render_definition_template(definition)
+    result = render_definition_template(definition, {}).project_context
 
     assert result == {
-        "definition_version": "1.1",
-        "env": {
-            "test_env": "    - app/*\n    - src\n",
-        },
+        "ctx": {
+            "definition_version": "1.1",
+            "env": ProjectEnvironment(
+                {
+                    "test_env": "    - app/*\n    - src\n",
+                },
+                {},
+            ),
+        }
     }
 
-    assert result["env"]["var_with_yml"] == "    - app/*\n    - src\n"
+    assert result["ctx"]["env"]["var_with_yml"] == "    - app/*\n    - src\n"
 
 
 @pytest.mark.parametrize(
@@ -357,7 +397,7 @@ def test_invalid_templating_syntax(template_value):
         },
     }
     with pytest.raises(InvalidTemplate) as err:
-        render_definition_template(definition)
+        render_definition_template(definition, {})
 
     assert err.value.message == f"Unexpected templating syntax in {template_value}"
 
@@ -368,7 +408,7 @@ def test_invalid_type_for_env_section():
         "env": ["test_env", "array_val1"],
     }
     with pytest.raises(InvalidTemplate) as err:
-        render_definition_template(definition)
+        render_definition_template(definition, {})
 
     assert (
         err.value.message
@@ -384,9 +424,130 @@ def test_invalid_type_for_env_variable():
         },
     }
     with pytest.raises(InvalidTemplate) as err:
-        render_definition_template(definition)
+        render_definition_template(definition, {})
 
     assert (
         err.value.message
-        == "Variable test_env in env section or project definition file should be a scalar"
+        == "Variable test_env in env section of project definition file should be a scalar"
     )
+
+
+@mock.patch.dict(os.environ, {"env_var_test": "value_from_os_env"}, clear=True)
+def test_env_priority_from_cli_and_os_env_and_project_env():
+    definition = {
+        "definition_version": "1.1",
+        "env": {
+            "env_var_test": "value_from_definition_file",
+            "final_value": "<% ctx.env.env_var_test %>",
+        },
+    }
+    result = render_definition_template(
+        definition, {"ctx": {"env": {"env_var_test": "value_from_cli_override"}}}
+    ).project_context
+
+    assert result == {
+        "ctx": {
+            "definition_version": "1.1",
+            "env": ProjectEnvironment(
+                {
+                    "env_var_test": "value_from_definition_file",
+                    "final_value": "value_from_cli_override",
+                },
+                {"env_var_test": "value_from_cli_override"},
+            ),
+        }
+    }
+
+    assert result["ctx"]["env"]["env_var_test"] == "value_from_cli_override"
+
+
+@mock.patch.dict(os.environ, {}, clear=True)
+def test_values_env_from_only_overrides():
+    definition = {
+        "definition_version": "1.1",
+        "env": {
+            "final_value": "<% ctx.env.env_var_test %>",
+        },
+    }
+    result = render_definition_template(
+        definition, {"ctx": {"env": {"env_var_test": "value_from_cli_override"}}}
+    ).project_context
+
+    assert result == {
+        "ctx": {
+            "definition_version": "1.1",
+            "env": ProjectEnvironment(
+                {"final_value": "value_from_cli_override"},
+                {"env_var_test": "value_from_cli_override"},
+            ),
+        }
+    }
+
+    assert result["ctx"]["env"]["env_var_test"] == "value_from_cli_override"
+
+
+@mock.patch.dict(os.environ, {}, clear=True)
+def test_cli_env_var_blank():
+    definition = {
+        "definition_version": "1.1",
+    }
+    result = render_definition_template(
+        definition, {"ctx": {"env": {"env_var_test": ""}}}
+    ).project_context
+
+    assert result == {
+        "ctx": {
+            "definition_version": "1.1",
+            "env": ProjectEnvironment(
+                {},
+                {"env_var_test": ""},
+            ),
+        }
+    }
+
+    assert result["ctx"]["env"]["env_var_test"] == ""
+
+
+@mock.patch.dict(os.environ, {}, clear=True)
+def test_cli_env_var_does_not_expand_with_templating():
+    definition = {
+        "definition_version": "1.1",
+    }
+    result = render_definition_template(
+        definition, {"ctx": {"env": {"env_var_test": "<% ctx.env.something %>"}}}
+    ).project_context
+
+    assert result == {
+        "ctx": {
+            "definition_version": "1.1",
+            "env": ProjectEnvironment(
+                {},
+                {"env_var_test": "<% ctx.env.something %>"},
+            ),
+        }
+    }
+
+    assert result["ctx"]["env"]["env_var_test"] == "<% ctx.env.something %>"
+
+
+@mock.patch.dict(os.environ, {"os_env_var": "os_env_var_value"}, clear=True)
+def test_os_env_and_override_envs_in_version_1():
+    definition = {
+        "definition_version": "1",
+    }
+
+    override_ctx = {"ctx": {"env": {"override_env": "override_env_value"}}}
+    result = render_definition_template(definition, override_ctx).project_context
+
+    assert result == {
+        "ctx": {
+            "definition_version": "1",
+            "env": ProjectEnvironment(
+                {},
+                {"override_env": "override_env_value"},
+            ),
+        }
+    }
+
+    assert result["ctx"]["env"]["override_env"] == "override_env_value"
+    assert result["ctx"]["env"]["os_env_var"] == "os_env_var_value"